requirements.txt
Django==4.2.5
mysqlclient==2.2.0
redis==5.0.0
djangorestframework==3.14.0
pandas==2.1.2
python-slugify==8.0.1
gunicorn==21.2.0
whitenoise==6.6.0
celery==5.3.4
hiredis==2.2.3
django-celery-beat==2.5.0
django-celery-results==2.5.1
djangorestframework-api-key==3.0.0